Moved user and users.js to the models directory
