+++ controllers/apiController.js
@@ -2,8 +2,8 @@ const path = require('path');
 const setup = require('../setup.js');
 const cache = require('../cache.js')(setup);
 const fleets = require('../models/fleets.js')(setup);
-const user = require('../user.js')(setup);
-const users = require('../users.js')(setup);
+const user = require('../models/user.js')(setup);
+const users = require('../models/users.js')(setup);
 const refresh = require('passport-oauth2-refresh');
 const banner = require('../waitlistBanner.js')(setup);
 const log = require('../logger.js')(module);
