+++ controllers/fcToolsController.js
@@ -2,7 +2,7 @@ var path = require('path');
 var fs = require('fs');
 var esi = require('eve-swagger');
 var setup = require('../setup.js');
-var users = require('../users.js')(setup);
+var users = require('../models/users.js')(setup);
 var refresh = require('passport-oauth2-refresh');
 const log = require('../logger.js')(module);
 const wlog = require('../wlog.js');
