+++ index.js
@@ -34,7 +34,7 @@ database.connect(function () {
 
 	//Custom imports
 
-	const users = require('./users.js')(setup);
+	const users = require('./models/users.js')(setup);
 	const customSSO = require('./customSSO.js')(refresh, setup, request, url);
 	const fleets = require('./models/fleets.js')(setup);
 	const waitlist = require('./models/globalWaitlist.js')(setup);
