+++ models/fleets.js
@@ -1,7 +1,7 @@
 const esi = require('eve-swagger');
 const refresh = require('passport-oauth2-refresh');
 const setup = require('../setup.js');
-const user = require('../user.js')(setup);
+const user = require('./user.js')(setup);
 const cache = require('../cache.js')(setup);
 const db = require('../dbHandler.js').db.collection('fleets');
 const log = require('../logger.js')(module);
