+++ models/user.js
@@ -1,10 +1,9 @@
-const fs = require('fs');
-const setup = require('./setup.js');
+const setup = require('../setup.js');
 const refresh = require('passport-oauth2-refresh');
 const esi = require('eve-swagger');
-const cache = require('./cache.js')(setup);
-const db = require('./dbHandler.js').db.collection('users');
-const log = require('./logger.js')(module);
+const cache = require('../cache.js')(setup);
+const db = require('../dbHandler.js').db.collection('users');
+const log = require('../logger.js')(module);
 
 module.exports = function() {
     /*
