+++ waitlistBanner.js
@@ -3,8 +3,6 @@ const path = require('path');
 const db = require('./dbHandler.js').db.collection('waitlist-banner');
 const ObjectId = require('mongodb').ObjectID;
 const setup = require('./setup.js');
-const user = require('./user.js')(setup);
-const users = require('./users.js')(setup);
 const log = require('./logger.js')(module);
 
 module.exports = function (setup) {
